tests/requirements-test.txt
requests==2.31.0
pytest==7.4.2
pytest-cov==4.1.0
python-dotenv==1.0.0
colorama==0.4.6
